add test for change owner
